Various Changes to the 3 pages, Oct. 25, 2018
Added yellow hover-overs to the header links, made the copyright in the footer into a link to my business website, also with a yellow hover-over. Discovered how to do drop-shadows on both text and objects, and added them to header and footer link type and header and footer background. Revised the embed code on the video so that it should pass validation. Added hover color of red to links on the home page and resume page. Still working on the media query.